# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис-шлюз для гарантированной доставки данных из AIS в 1С через персистентную очередь.",
        "title": "AIS-1C Integration Proxy API",
        "contact": {
            "name": "Renat"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/v1/data": {
            "put": {
                "description": "Принимает JSON пакет от AIS, ставит в очередь на отправку в 1С.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integration"
                ],
                "summary": "Прием данных (Sale/Update/Delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Пакет данных",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AISRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Принимает JSON пакет от AIS, ставит в очередь на отправку в 1С.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integration"
                ],
                "summary": "Прием данных (Sale/Update/Delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Пакет данных",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AISRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Принимает JSON пакет от AIS, ставит в очередь на отправку в 1С.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Integration"
                ],
                "summary": "Прием данных (Sale/Update/Delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Пакет данных",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AISRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/v1/health": {
            "get": {
                "description": "Проверяет доступность сервиса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AISDocument": {
            "type": "object",
            "properties": {
                "ClientActualAddress": {
                    "type": "string"
                },
                "ClientCity": {
                    "type": "string"
                },
                "ClientClientTypeId": {
                    "description": "Enum: 1=Legal, 2=Resident, 3=NonResident",
                    "type": "string"
                },
                "ClientContractorTypeId": {
                    "type": "string"
                },
                "ClientEmail": {
                    "type": "string"
                },
                "ClientFieldOfActivity": {
                    "type": "string"
                },
                "ClientFullName": {
                    "type": "string"
                },
                "ClientId": {
                    "description": "--- Client Fields ---\nRequired: Id, Type, Voen, Name. Others are optional.",
                    "type": "string"
                },
                "ClientIndex": {
                    "type": "string"
                },
                "ClientLeaderDuty": {
                    "type": "string"
                },
                "ClientLeaderFullName": {
                    "type": "string"
                },
                "ClientLegalAddress": {
                    "type": "string"
                },
                "ClientPhone": {
                    "type": "string"
                },
                "ClientVOENOrPassportNo": {
                    "type": "string"
                },
                "ClientWebsite": {
                    "type": "string"
                },
                "ContractEndDate": {
                    "description": "Optional",
                    "type": "string"
                },
                "ContractId": {
                    "description": "--- Contract Fields ---\nAll required except EndDate",
                    "type": "string"
                },
                "ContractIsTransfer": {
                    "type": "boolean"
                },
                "ContractNo": {
                    "type": "string"
                },
                "ContractPayTime": {
                    "type": "string"
                },
                "ContractSignedDate": {
                    "type": "string"
                },
                "ContractStartDate": {
                    "type": "string"
                },
                "ContractStatus": {
                    "type": "string"
                },
                "ContractType": {
                    "type": "string"
                },
                "DepartmentName": {
                    "type": "string"
                },
                "PayStatusName": {
                    "type": "string"
                },
                "PrecinctName": {
                    "description": "--- Reference Fields ---",
                    "type": "string"
                },
                "Privilege": {
                    "type": "string"
                },
                "SaleCarNumber": {
                    "type": "string"
                },
                "SaleClientId": {
                    "type": "string"
                },
                "SaleComment": {
                    "description": "Required for Cancellation (DELETE)",
                    "type": "string"
                },
                "SaleDate": {
                    "description": "Format: ISO8601 or YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "SaleDeclarationNo": {
                    "type": "string"
                },
                "SaleDetails": {
                    "description": "--- Array of Nomenclatures ---",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SaleDetail"
                    }
                },
                "SaleId": {
                    "description": "--- Sale Fields ---",
                    "type": "string"
                },
                "SaleInspectorId": {
                    "type": "string"
                },
                "SaleIsDiplomatic": {
                    "type": "boolean"
                },
                "SaleMainAmount": {
                    "type": "number"
                },
                "SalePayDate": {
                    "type": "string"
                },
                "SalePayStatusId": {
                    "description": "Enum: 1=NotPaid, 2=Paid, 3=Partial, 4=Cancelled, 5=Pending, 6=Return, etc.",
                    "type": "string"
                },
                "SalePayType": {
                    "type": "string"
                },
                "SalePaymentType": {
                    "description": "Enum: 0=None, 1=Avans, 3=Hop, 4=Vais, 5=Contract, 6=Manual, 7=Emanat, 8=Kocurme",
                    "type": "string"
                },
                "SalePrecinctId": {
                    "type": "string"
                },
                "SaleVatAmount": {
                    "type": "number"
                },
                "UserActive": {
                    "type": "boolean"
                },
                "UserEmail": {
                    "type": "string"
                },
                "UserFirstName": {
                    "type": "string"
                },
                "UserId": {
                    "description": "--- User Fields ---\nRequired for Cancellation: UserId",
                    "type": "string"
                },
                "UserLastName": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                },
                "UserPhone": {
                    "type": "string"
                },
                "UserPrecinctId": {
                    "type": "string"
                },
                "UserRegistrationDate": {
                    "type": "string"
                },
                "UserStatusId": {
                    "description": "Enum: 1=SuperAdmin, 2=Accountant, 3=Inspector, ...",
                    "type": "string"
                }
            }
        },
        "models.AISRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.AISDocument"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "sale, sale_update, sale_cancellation, etc.",
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.SaleDetail": {
            "type": "object",
            "properties": {
                "SaleDetailAmount": {
                    "type": "number"
                },
                "SaleDetailId": {
                    "type": "string"
                },
                "SaleDetailPrice": {
                    "type": "number"
                },
                "SaleDetailQuantity": {
                    "type": "number"
                },
                "SaleDetailSaleId": {
                    "type": "string"
                },
                "SaleDetailServiceId": {
                    "type": "string"
                },
                "SaleDetailTotalAmount": {
                    "type": "number"
                },
                "SaleDetailVATAmount": {
                    "type": "number"
                },
                "SaleDetailVATRate": {
                    "type": "number"
                },
                "ServiceCode": {
                    "type": "string"
                },
                "ServiceName": {
                    "type": "string"
                },
                "ServiceUnit": {
                    "type": "string"
                }
            }
        }
    }
}